requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
anthropic>=0.90.0
google-genai>=1.0.0
pydantic>=2.0
python-dotenv>=1.0.0
pyyaml>=6.0
